refactor(ui): improves `Matrix` folder and files overall organization
