+++ webapp/src/components/App/Singlestudy/explore/Modelization/Areas/Hydro/Correlation/utils.ts
@@ -14,7 +14,7 @@
 
 import { StudyMetadata, Area } from "../../../../../../../../common/types";
 import client from "../../../../../../../../services/api/client";
-import { MatrixDataDTO } from "../../../../../../../common/MatrixGrid/types";
+import { MatrixDataDTO } from "../../../../../../../common/Matrix/core/types";
 import { AreaCoefficientItem } from "../utils";
 
 ////////////////////////////////////////////////////////////////
