+++ webapp/src/components/App/Singlestudy/explore/Modelization/Areas/Reserve.tsx
@@ -14,7 +14,7 @@
 
 import useAppSelector from "../../../../../../redux/hooks/useAppSelector";
 import { getCurrentAreaId } from "../../../../../../redux/selectors";
-import Matrix from "../../../../../common/MatrixGrid/Matrix";
+import Matrix from "../../../../../common/Matrix";
 
 function Reserve() {
   const currentArea = useAppSelector(getCurrentAreaId);
