+++ webapp/src/components/common/Matrix/components/MatrixActions/MatrixActions.test.tsx
@@ -15,7 +15,7 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import MatrixActions from "./MatrixActions";
+import MatrixActions from ".";
 
 vi.mock("../buttons/SplitButton", () => ({
   default: ({
