+++ webapp/src/components/common/Matrix/components/MatrixActions/index.tsx
@@ -13,13 +13,11 @@
  */
 
 import { Box, Divider, IconButton, Tooltip } from "@mui/material";
-import SplitButton from "../buttons/SplitButton";
-import FileDownload from "@mui/icons-material/FileDownload";
+import SplitButton from "@/components/common/buttons/SplitButton";
+import DownloadMatrixButton from "@/components/common/buttons/DownloadMatrixButton";
+import { FileDownload, Save, Undo, Redo } from "@mui/icons-material";
 import { useTranslation } from "react-i18next";
 import { LoadingButton } from "@mui/lab";
-import Save from "@mui/icons-material/Save";
-import { Undo, Redo } from "@mui/icons-material";
-import DownloadMatrixButton from "../buttons/DownloadMatrixButton";
 
 interface MatrixActionsProps {
   onImport: VoidFunction;
